+++ core-processor/src/main/java/jeff/core/controller/CoreController.java
@@ -4,10 +4,12 @@ import com.fasterxml.jackson.databind.JsonNode;
 import jeff.common.entity.dto.send.ResponseObject;
 import jeff.core.exception.OrderException;
 import jeff.core.service.OrderService;
+import org.springframework.http.HttpRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
 
 /**
  * 本伺服端的入口控制器。
@@ -25,8 +27,8 @@ public class CoreController {
      * 通常是購物車結帳後按下送出所打的API。
      */
     @PostMapping("/order/normal")
-    public ResponseEntity<ResponseObject> createAnOrderFromNormalSalesEvent(@RequestBody JsonNode param) throws OrderException {
-        return ResponseEntity.ok(normalOrderService.createOrder(param));
+    public ResponseEntity<ResponseObject> createAnOrderFromNormalSalesEvent(@RequestBody JsonNode param, @RequestAttribute String UUID) throws OrderException {
+        return ResponseEntity.ok(normalOrderService.createOrder(param, UUID));
     }
 
 }