+++ my-double-datasource-api/src/main/java/jeff/persistent/model/mongo/dao/FlashSaleEventLogRepo.java
@@ -1,10 +1,10 @@
 package jeff.persistent.model.mongo.dao;
 
-import jeff.persistent.model.mongo.po.FlashSaleTempRecord;
+import jeff.persistent.model.mongo.po.FlashSaleEventLog;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FlashSaleTempRecordRepo extends MongoRepository<FlashSaleTempRecord, String> {
+public interface FlashSaleEventLogRepo extends MongoRepository<FlashSaleEventLog, String> {
 
 }